Filter out dot files
This allows me to manage my notes using source control, git/mercurial

We skip files that either start with a '.' or the last portion after
the last '\' start with a '.'
When searching grep ignores these anyway so we don't need to filter
what is returned there.

Note: This is probably not Windows safe. If you want it to work on
Win then you'll have to fix it yourself.